moved general hubsot api initialization to base
